+++ packages/client-bonsai/src/utils/ipfs.ts
@@ -71,49 +71,43 @@ const pinFile = async (file: {
   return storjGatewayURL(response.data.Hash);
 };
 
-export const parseAndUploadBase64Image = async (imageResponse) => {
-  if (imageResponse.success && imageResponse.data?.[0]) {
-    if (!imageResponse.data[0].includes("base64")) {
-      console.error("Image response does not contain base64 data");
-      return;
-    }
-    // Convert base64 to buffer
-    const base64Data = imageResponse.data[0].replace(
-      /^data:image\/\w+;base64,/,
-      ""
-    );
-    const imageBuffer = Buffer.from(base64Data, "base64");
-
-    // Create a file object that can be used with FormData
-    const file = {
-      buffer: imageBuffer,
-      originalname: `generated_${Date.now()}.png`,
-      mimetype: "image/png",
-    };
-
-    // Upload to your hosting service
-    return await pinFile(file);
+export const parseAndUploadBase64Image = async (imageDataBase64: string) => {
+  if (!imageDataBase64.includes("base64")) {
+    console.error("Image response does not contain base64 data");
+    return;
   }
+  // Convert base64 to buffer
+  const base64Data = imageDataBase64.replace(
+    /^data:image\/\w+;base64,/,
+    ""
+  );
+  const imageBuffer = Buffer.from(base64Data, "base64");
 
-  console.error("No image response");
+  // Create a file object that can be used with FormData
+  const file = {
+    buffer: imageBuffer,
+    originalname: `generated_${Date.now()}.png`,
+    mimetype: "image/png",
+  };
+
+  // Upload to your hosting service
+  return await pinFile(file);
 };
 
-export const parseBase64Image = (imageResponse): File | undefined => {
-  if (imageResponse.success && imageResponse.data?.[0]) {
-    if (!imageResponse.data[0].includes("base64")) {
-      console.error("Image response does not contain base64 data");
-      return;
-    }
-    // Convert base64 to buffer
-    const base64Data = imageResponse.data[0].replace(
-      /^data:image\/\w+;base64,/,
-      ""
-    );
-    const imageBuffer = Buffer.from(base64Data, "base64");
-
-    // Create a file object that can be used with FormData
-    return new File([imageBuffer], `generated_${Date.now()}.png`, {
-      type: "image/png"
-    });
+export const parseBase64Image = (imageDataBase64: string): File | undefined => {
+  if (!imageDataBase64.includes("base64")) {
+    console.error("Image response does not contain base64 data");
+    return;
   }
+  // Convert base64 to buffer
+  const base64Data = imageDataBase64.replace(
+    /^data:image\/\w+;base64,/,
+    ""
+  );
+  const imageBuffer = Buffer.from(base64Data, "base64");
+
+  // Create a file object that can be used with FormData
+  return new File([imageBuffer], `generated_${Date.now()}.png`, {
+    type: "image/png"
+  });
 };
